requirements.txt
backports.zoneinfo;python_version<"3.9"
blinker==1.7.0
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dateparser==1.1.8
dnspython==2.4.2
flask==3.0.0
Flask-PyMongo==2.3.0
importlib-metadata==6.8.0
invoice2data==0.4.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pycparser==2.21
pymongo==4.6.0
pypdf2==3.0.1
pypdfium2==4.23.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
six==1.16.0
typing-extensions==4.8.0
tzlocal==5.2
werkzeug==3.0.1
zipp==3.17.0
